Split resolver declaration validation
