+++ docs/COMPLETION_AUDIT.md
@@ -2747,6 +2747,9 @@ and do not assume Phase 4 is ready without evidence.
   `src/parser/core.rs` and `src/parser/precedence.rs`, preserving parser unit
   coverage while keeping `src/parser/mod.rs` focused on module wiring and the
   public parse entry point.
+- Resolver declaration validation dispatch now lives in
+  `src/resolver/declaration_validation.rs`, preserving resolver Phase 2
+  coverage while keeping `src/resolver.rs` focused on top-level orchestration.
 - Resolver declaration replay-kind helpers now live in
   `src/typechecker/declaration_collection_resolver_tasks/replay_kinds.rs`,
   preserving type, behavior, impl, callable, and type-reference replay task
